# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "title": "IsoGate prototype.",
    "version": "1.0"
  },
  "paths": {
...
    "/subject": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "summary": "Get a specific subject by it's UUID.",
        "operationId": "GetSubject",
        "responses": {
          "200": {
            "$ref": "#/responses/SubjectResponse"
          },
          "401": {
            "$ref": "#/responses/HttpResponse"
          }
        }
      },
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "summary": "Creates or updates a subject record.",
        "operationId": "SaveSubject",
        "responses": {
          "200": {
            "$ref": "#/responses/SubjectResponse"
          },
          "401": {
            "$ref": "#/responses/HttpResponse"
          },
          "409": {
            "$ref": "#/responses/SubjectResponse"
          }
        }
      }
    },
    "/subjects": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "summary": "Gets all subjects for a given Site UUID.",
        "operationId": "GetSubjects",
        "responses": {
          "200": {
            "$ref": "#/responses/SubjectsResponse"
          },
          "401": {
            "$ref": "#/responses/HttpResponse"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Subject": {
      "description": "Subject model",
      "type": "object",
      "properties": {
        "localUpdateDateTime": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "LocalUpdateDateTime"
        },
        "serverUpdateDateTime": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "ServerUpdateDateTime"
        },
        "subjectActive": {
          "type": "boolean",
          "x-go-name": "SubjectActive"
        },
        "subjectAdditionalId": {
          "type": "string",
          "x-go-name": "SubjectAdditionalID"
        },
        "subjectAdditionalId2": {
          "type": "string",
          "x-go-name": "SubjectAdditionalID2"
        },
        "subjectAddress": {
          "type": "string",
          "x-go-name": "SubjectAddress"
        },
        "subjectCreationDate": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "SubjectCreationDate"
        },
        "subjectDob": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "SubjectDob"
        },
        "subjectEmail": {
          "type": "string",
          "x-go-name": "SubjectEmail"
        },
        "subjectEthnicity": {
          "type": "string",
          "x-go-name": "SubjectEthnicity"
        },
        "subjectFirstName": {
          "type": "string",
          "x-go-name": "SubjectFirstName"
        },
        "subjectGender": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "SubjectGender"
        },
        "subjectGroupUuid": {
          "type": "string",
          "x-go-name": "SubjectGroupUUID"
        },
        "subjectHeight": {
          "type": "number",
          "format": "float",
          "x-go-name": "SubjectHeight"
        },
        "subjectId": {
          "type": "string",
          "x-go-name": "SubjectID"
        },
        "subjectLastName": {
          "type": "string",
          "x-go-name": "SubjectLastName"
        },
        "subjectPhoneNumber": {
          "type": "string",
          "x-go-name": "SubjectPhoneNumber"
        },
        "subjectSiteUuid": {
          "type": "string",
          "x-go-name": "SubjectSiteUUID"
        },
        "subjectStatus": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "SubjectStatus"
        },
        "subjectUserName": {
          "type": "string",
          "x-go-name": "SubjectUserName"
        },
        "subjectUuid": {
          "type": "string",
          "x-go-name": "SubjectUUID"
        }
      },
      "x-go-package": "isogate/pkg/models"
    }
  },
  "responses": {
...
    "HttpResponse": {
      "description": "An ok http response"
    },
...
    "SubjectResponse": {
      "description": "Subject model",
      "headers": {
        "localUpdateDateTime": {
          "type": "string",
          "format": "date-time"
        },
        "serverUpdateDateTime": {
          "type": "string",
          "format": "date-time"
        },
        "subjectActive": {
          "type": "boolean"
        },
        "subjectAdditionalId": {
          "type": "string"
        },
        "subjectAdditionalId2": {
          "type": "string"
        },
        "subjectAddress": {
          "type": "string"
        },
        "subjectCreationDate": {
          "type": "string",
          "format": "date-time"
        },
        "subjectDob": {
          "type": "string",
          "format": "date-time"
        },
        "subjectEmail": {
          "type": "string"
        },
        "subjectEthnicity": {
          "type": "string"
        },
        "subjectFirstName": {
          "type": "string"
        },
        "subjectGender": {
          "type": "integer",
          "format": "int64"
        },
        "subjectGroupUuid": {
          "type": "string"
        },
        "subjectHeight": {
          "type": "number",
          "format": "float"
        },
        "subjectId": {
          "type": "string"
        },
        "subjectLastName": {
          "type": "string"
        },
        "subjectPhoneNumber": {
          "type": "string"
        },
        "subjectSiteUuid": {
          "type": "string"
        },
        "subjectStatus": {
          "type": "integer",
          "format": "int64"
        },
        "subjectUserName": {
          "type": "string"
        },
        "subjectUuid": {
          "type": "string"
        }
      }
    },
    "SubjectsResponse": {
      "description": "Subjects is a collection of Subjects",
      "schema": {
        "type": "array",
        "items": {
          "$ref": "#/definitions/Subject"
        }
      }
    }
  }
}